+++ MyToDo/ViewModels/ToDoViewModel.cs
@@ -1,4 +1,6 @@
 ﻿using Microsoft.Win32;
+using MyToDo.Common;
+using MyToDo.Extensions;
 using MyToDo.Service;
 using MyToDo.Shared.Dtos;
 using MyToDo.Shared.Parameters;
@@ -17,6 +19,9 @@ namespace MyToDo.ViewModels
 {
     public class ToDoViewModel : NavigationViewModel
     {
+
+        private readonly IDialogHostService dialogHost;
+
         public ToDoViewModel(IToDoService service, IContainerProvider provider)
             : base(provider)
         {
@@ -24,6 +29,7 @@ namespace MyToDo.ViewModels
             ExecuteCommand = new DelegateCommand<string>(Execute);
             SelectedCommand = new DelegateCommand<ToDoDto>(Selected);
             DeleteCommand = new DelegateCommand<ToDoDto>(Delete);
+            dialogHost = provider.Resolve<IDialogHostService>();    // 从Prism的Ioc容器中获取出来（IDialogService的父类，所以可以取出来）
             this.service = service;
         }
 
@@ -31,6 +37,13 @@ namespace MyToDo.ViewModels
         {
             try
             {
+                // 这里调用扩展方法
+                // 删除需要确认
+                var dialogResult = await dialogHost.Question("温馨提示", $"确认删除待办事项:{obj.Title}");
+
+                if (dialogResult.Result != Prism.Services.Dialogs.ButtonResult.OK) return;
+
+
                 UpdateLoading(true);
                 // 删除数据库的
                 var deleteResult = await service.DeleteAsync(obj.Id);
